added docker

+++ ToDoService.Application/Handlers/DeleteToDoHandler.cs
@@ -13,8 +13,20 @@ public class DeleteToDoHandler : IRequestHandler<DeleteToDoCommand, bool>
         _repository = repository;
     }
 
-public async Task<bool> Handle(DeleteToDoCommand request, CancellationToken cancellationToken)
+    public async Task<bool> Handle(DeleteToDoCommand request, CancellationToken cancellationToken)
     {
+        var todo = await _repository.GetByIdAsync(request.Id);
+        
+        if (todo == null)
+        {
+            return false;
+        }
+
+        if (todo.UserId != request.UserId)
+        {
+            throw new UnauthorizedAccessException("You don't have permission to delete this item");
+        }
+
         return await _repository.DeleteAsync(request.Id);
     }
 }
